# openapi.yaml
openapi: 3.0.0
info:
  title: Another Home API
  description: >
    Unified API contract for the Another Home Hostel Management System, aggregating the
    Accommodation, Finance, and Gateway (Admin) microservices behind the API Gateway.


    All routes are reached through the Gateway at `/api/v1/*`. The Gateway authenticates every
    request against Asgardeo (see the `bearerAuth` security scheme) before proxying it to the
    owning microservice, injecting `x-user-id` and `x-user-roles` headers for the service to read.


    There is no local `/auth/*` endpoint — login/registration is handled entirely by Asgardeo
    (hosted login / OIDC), which issues the JWT clients send as the Bearer token.


    The **Operations** and **Notifications** sections at the bottom of this document describe
    planned microservices that have not been implemented yet — they exist for design reference
    only and are not reachable through the Gateway.
  version: '2.0'
servers:
  - url: http://localhost:3001
    description: API Gateway (local docker-compose)
  - url: http://localhost:4001
    description: Accommodation service (direct, dev only — bypasses Gateway auth)
  - url: http://localhost:4003
    description: Finance service (direct, dev only — bypasses Gateway auth)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/accommodation/rooms:
    get:
      tags: [Accommodation]
      summary: Get all rooms and their current capacity
      responses:
        '200':
          description: Returns an array of all rooms.
    post:
      tags: [Accommodation]
      summary: Create a new hostel room
      description: 'Requires role: staff, warden, or super-admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomDto'
      responses:
        '201':
          description: Room created successfully.
  /api/v1/accommodation/rooms/{id}:
    patch:
      tags: [Accommodation]
      summary: Update an existing hostel room
      description: 'Requires role: staff, warden, or super-admin.'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomDto'
      responses:
        '200':
          description: Room updated successfully.
    delete:
      tags: [Accommodation]
      summary: Delete a hostel room
      description: 'Requires role: staff, warden, or super-admin.'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Room deleted successfully.
  /api/v1/accommodation/rooms/{id}/assign:
    post:
      tags: [Accommodation]
      summary: Assign a student to the next available bed in a room
      description: 'Requires role: staff, warden, or super-admin.'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoomDto'
      responses:
        '201':
          description: Student assigned to room successfully.
  /api/v1/accommodation/beds:
    post:
      tags: [Accommodation]
      summary: Create a new bed in a room
      description: 'Requires role: staff, warden, or super-admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBedDto'
      responses:
        '201':
          description: Bed created successfully.
  /api/v1/accommodation/allocations:
    get:
      tags: [Accommodation]
      summary: Get a list of all current bed allocations
      responses:
        '200':
          description: Returns all occupied beds.
    post:
      tags: [Accommodation]
      summary: Assign a student to a specific bed
      description: 'Requires role: staff, warden, or super-admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocateBedDto'
      responses:
        '201':
          description: Student successfully allocated.
        '400':
          description: Bed already occupied or not found.
  /api/v1/accommodation/buildings:
    get:
      tags: [Accommodation]
      summary: Get all buildings with occupancy summary
      responses:
        '200':
          description: Returns an array of all buildings.
    post:
      tags: [Accommodation]
      summary: Create a new hostel building
      description: 'Requires role: staff, warden, or super-admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingDto'
      responses:
        '201':
          description: Building created successfully.
  /api/v1/accommodation/students:
    get:
      tags: [Accommodation]
      summary: Get all students with their current room assignment
      responses:
        '200':
          description: Returns an array of all students.
    post:
      tags: [Accommodation]
      summary: Register a new student
      description: 'Requires role: staff, warden, or super-admin.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentDto'
      responses:
        '201':
          description: Student created successfully.

  # ==========================================
  # 2. FINANCE (Finance Microservice)
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Asgardeo-issued access token. The Gateway verifies it against the Asgardeo JWKS before proxying the request.
  schemas:
...
    CreateRoomDto:
      type: object
      required:
        - roomNumber
        - capacity
        - gender
        - airConditioning
        - rentPerMonth
        - floor
      properties:
        roomNumber:
          type: string
          example: A-101
          description: The unique identifier for the room (e.g., Block-Number).
        capacity:
          type: integer
          example: 4
          description: Maximum number of beds in this room.
        gender:
          type: string
          enum: [Male, Female, Neutral]
          example: Male
          description: The designated gender for this specific room or block.
        airConditioning:
          type: string
          enum: [AC, Non-AC]
          example: Non-AC
        rentPerMonth:
          type: number
          example: 5000
          description: Monthly rent for the room in local currency.
        floor:
          type: integer
          example: 1
        buildingId:
          type: string
          nullable: true
          description: The ID of the building this room belongs to.
    UpdateRoomDto:
      type: object
      description: All fields are optional — only the fields provided are updated.
      properties:
        roomNumber:
          type: string
        capacity:
          type: integer
        gender:
          type: string
          enum: [Male, Female, Neutral]
        airConditioning:
          type: string
          enum: [AC, Non-AC]
        rentPerMonth:
          type: number
        floor:
          type: integer
        buildingId:
          type: string
    AssignRoomDto:
      type: object
      required:
        - studentId
      properties:
        studentId:
          type: string
          example: STU-2023-045
          description: The ID of the student being assigned to the room.
    CreateBedDto:
      type: object
      required:
        - bedId
        - bedNumber
        - roomId
      properties:
        bedId:
          type: string
          description: The unique ID of the bed.
        bedNumber:
          type: string
          description: The display number or name of the bed.
        roomId:
          type: string
          description: The ID of the room this bed belongs to.
    AllocateBedDto:
      type: object
      required:
        - studentId
        - bedId
      properties:
        studentId:
          type: string
          example: STU-2023-045
        bedId:
          type: string
          example: BED-A101-01
    CreateBuildingDto:
      type: object
      required:
        - name
        - address
        - floorCount
      properties:
        name:
          type: string
          example: Sunrise Block
        address:
          type: string
          example: 12 Lake Road, Colombo
        floorCount:
          type: integer
          example: 4
    CreateStudentDto:
      type: object
      required:
        - studentCode
        - name
        - email
        - contact
      properties:
        studentCode:
          type: string
          example: AH-1000
        name:
          type: string
          example: Ishara Perera
        email:
          type: string
          format: email
          example: ishara.perera@student.anotherhome.com
        contact:
          type: string
          example: '+94 71 234 5678'
        guardianName:
          type: string
        guardianContact:
          type: string
        address:
          type: string

    # --- Finance Schemas ---
...